% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_stacker.R
\name{Auto_stacker}
\alias{Auto_stacker}
\title{Stacks all rasters within a folder}
\usage{
Auto_stacker(
  path,
  file_type = 1,
  projection_CRS = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{path}{The Path to the folder containing the .h5 files}

\item{file_type}{Default is 1. 1. DNB_BRDF-Corrected_NTL 2. DNB_Lunar_Irradiance 3. Gap_Filled_DNB_BRDF-Corrected_NTL 4. Latest_High_Quality_Retrieval 5. Mandatory_Quality_Flag 6. QF_Cloud_Mask 7. Snow_Flag}

\item{projection_CRS}{The CRS projection. The default is "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"}
}
\value{
A stack of rasters
}
\description{
Stacks all rasters within a folder
}
