% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5_data_extracter.R
\name{H5_data_extracter}
\alias{H5_data_extracter}
\title{Coverts NASA h5 to rasters}
\usage{
H5_data_extracter(path, file_type = 1, file_name)
}
\arguments{
\item{path}{The Path to the folder containing the .h5 files}

\item{file_type}{Default is 1. 1. DNB_BRDF-Corrected_NTL 2. DNB_Lunar_Irradiance 3. Gap_Filled_DNB_BRDF-Corrected_NTL 4. Latest_High_Quality_Retrieval 5. Mandatory_Quality_Flag 6. QF_Cloud_Mask 7. Snow_Flag}

\item{file_name}{The name of the file to be extracted}
}
\value{
A stack of rasters
}
\description{
Coverts NASA h5 to rasters
}
