% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stack_Sigma.R
\name{Stack_sigma}
\alias{Stack_sigma}
\title{Calculates Sigma for a raster stack}
\usage{
Stack_sigma(x)
}
\arguments{
\item{x}{The raster stack}
}
\value{
The sigma
}
\description{
Calculates Sigma for a raster stack
}
