% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Welch_t_statistic.R
\name{Welch_t}
\alias{Welch_t}
\title{Estimates the Welch t-statistic for two raster stacks}
\usage{
Welch_t(stack1, stack2)
}
\arguments{
\item{stack1}{The base raster stack}

\item{stack2}{The raster stack to be compared with}
}
\value{
The Welch t-statistic for two raster stacks
}
\description{
Estimates the Welch t-statistic for two raster stacks
}
